+++ src/frontend/apps/impress/src/features/docs/doc-editor/components/BlockNoteEditor.tsx
@@ -26,7 +26,13 @@ const cssEditor = (readonly: boolean) => `
   };
   & .bn-editor {
     padding-right: 30px;
-    ${readonly && `padding-left: 30px;`}
+    ${
+      readonly &&
+      `
+        padding-left: 30px;
+        pointer-events: none;
+      `
+    }
   };
   & .collaboration-cursor__caret.ProseMirror-widget{
     word-wrap: initial;
@@ -139,14 +145,14 @@ export const BlockNoteContent = ({
         provider,
         fragment: provider.document.getXmlFragment('document-store'),
         user: {
-          name: userData?.email || 'Anonymous',
+          name: userData?.full_name || userData?.email || 'Anonymous',
           color: randomColor(),
         },
       },
       dictionary: locales[lang as keyof typeof locales],
       uploadFile,
     },
-    [provider, uploadFile, userData?.email, lang],
+    [lang, provider, uploadFile, userData?.email, userData?.full_name],
   );
 
   useEffect(() => {